requirements.txt
chromadb
pandas
requests
python-dotenv
